+++ tests/liquidateMaxLps.ts
@@ -1,0 +1,227 @@
+import * as anchor from '@coral-xyz/anchor';
+import {
+	BASE_PRECISION,
+	BN,
+	OracleSource,
+	OracleGuardRails,
+	TestClient,
+	EventSubscriber,
+	PRICE_PRECISION,
+	PositionDirection,
+	Wallet,
+	LIQUIDATION_PCT_PRECISION,
+} from '../sdk/src';
+
+import { Program } from '@coral-xyz/anchor';
+
+import { Keypair, PublicKey } from '@solana/web3.js';
+
+import {
+	mockOracle,
+	mockUSDCMint,
+	mockUserUSDCAccount,
+	setFeedPrice,
+	initializeQuoteSpotMarket,
+	printTxLogs,
+} from './testHelpers';
+import {
+	BulkAccountLoader,
+	findComputeUnitConsumption,
+	PERCENTAGE_PRECISION,
+} from '../sdk';
+
+describe('max lp liq', () => {
+	const provider = anchor.AnchorProvider.local(undefined, {
+		preflightCommitment: 'confirmed',
+		commitment: 'confirmed',
+	});
+	const connection = provider.connection;
+	anchor.setProvider(provider);
+	const chProgram = anchor.workspace.Drift as Program;
+
+	let driftClient: TestClient;
+	const eventSubscriber = new EventSubscriber(connection, chProgram, {
+		commitment: 'recent',
+	});
+	eventSubscriber.subscribe();
+
+	const bulkAccountLoader = new BulkAccountLoader(connection, 'confirmed', 1);
+
+	let usdcMint;
+	let userUSDCAccount;
+
+	const liquidatorKeyPair = new Keypair();
+	let liquidatorUSDCAccount: Keypair;
+	let liquidatorDriftClient: TestClient;
+
+	// ammInvariant == k == x * y
+	const mantissaSqrtScale = new BN(Math.sqrt(PRICE_PRECISION.toNumber()));
+	const ammInitialQuoteAssetReserve = new anchor.BN(5 * 10 ** 13).mul(
+		mantissaSqrtScale
+	);
+	const ammInitialBaseAssetReserve = new anchor.BN(5 * 10 ** 13).mul(
+		mantissaSqrtScale
+	);
+
+	const usdcAmount = new BN(10 * 10 ** 6);
+	const nLpShares = new BN(10000000);
+
+	let oracle: PublicKey;
+	const numMkts = 8;
+
+	before(async () => {
+		usdcMint = await mockUSDCMint(provider);
+		userUSDCAccount = await mockUserUSDCAccount(usdcMint, usdcAmount, provider);
+
+		oracle = await mockOracle(1);
+
+		driftClient = new TestClient({
+			connection,
+			wallet: provider.wallet,
+			programID: chProgram.programId,
+			opts: {
+				commitment: 'confirmed',
+			},
+			activeSubAccountId: 0,
+			perpMarketIndexes: [0],
+			spotMarketIndexes: [0],
+			oracleInfos: [
+				{
+					publicKey: oracle,
+					source: OracleSource.PYTH,
+				},
+			],
+			accountSubscription: {
+				type: 'polling',
+				accountLoader: bulkAccountLoader,
+			},
+		});
+
+		await driftClient.initialize(usdcMint.publicKey, true);
+		await driftClient.subscribe();
+
+		await driftClient.updateInitialPctToLiquidate(
+			LIQUIDATION_PCT_PRECISION.toNumber()
+		);
+
+		await initializeQuoteSpotMarket(driftClient, usdcMint.publicKey);
+		await driftClient.updatePerpAuctionDuration(new BN(0));
+
+		const periodicity = new BN(0);
+
+		for (let i = 0; i < numMkts; i++) {
+			await driftClient.initializePerpMarket(
+				i,
+				oracle,
+				ammInitialBaseAssetReserve,
+				ammInitialQuoteAssetReserve,
+				periodicity
+			);
+		}
+
+		await driftClient.initializeUserAccountAndDepositCollateral(
+			usdcAmount,
+			userUSDCAccount.publicKey
+		);
+
+		for (let i = 0; i < numMkts; i++) {
+			await driftClient.openPosition(
+				PositionDirection.LONG,
+				new BN(175)
+					.mul(BASE_PRECISION)
+					.div(new BN(10))
+					.divn(numMkts * 4),
+				i,
+				new BN(0)
+			);
+
+			const txSig = await driftClient.addPerpLpShares(
+				nLpShares.divn(numMkts * 4),
+				i
+			);
+			await printTxLogs(connection, txSig);
+		}
+
+		provider.connection.requestAirdrop(liquidatorKeyPair.publicKey, 10 ** 9);
+		liquidatorUSDCAccount = await mockUserUSDCAccount(
+			usdcMint,
+			usdcAmount,
+			provider,
+			liquidatorKeyPair.publicKey
+		);
+		liquidatorDriftClient = new TestClient({
+			connection,
+			wallet: new Wallet(liquidatorKeyPair),
+			programID: chProgram.programId,
+			opts: {
+				commitment: 'confirmed',
+			},
+			activeSubAccountId: 0,
+			perpMarketIndexes: [0, 1, 2, 3, 4, 5, 6, 7],
+			spotMarketIndexes: [0],
+			oracleInfos: [
+				{
+					publicKey: oracle,
+					source: OracleSource.PYTH,
+				},
+			],
+			accountSubscription: {
+				type: 'polling',
+				accountLoader: bulkAccountLoader,
+			},
+		});
+		await liquidatorDriftClient.subscribe();
+
+		await liquidatorDriftClient.initializeUserAccountAndDepositCollateral(
+			usdcAmount,
+			liquidatorUSDCAccount.publicKey
+		);
+	});
+
+	after(async () => {
+		await driftClient.unsubscribe();
+		await liquidatorDriftClient.unsubscribe();
+		await eventSubscriber.unsubscribe();
+	});
+
+	it('liquidate', async () => {
+		await setFeedPrice(anchor.workspace.Pyth, 0.1, oracle);
+
+		const oracleGuardRails: OracleGuardRails = {
+			priceDivergence: {
+				markOraclePercentDivergence: PERCENTAGE_PRECISION,
+				oracleTwap5MinPercentDivergence: PERCENTAGE_PRECISION.div(new BN(10)),
+			},
+			validity: {
+				slotsBeforeStaleForAmm: new BN(100),
+				slotsBeforeStaleForMargin: new BN(100),
+				confidenceIntervalMaxSize: new BN(100000),
+				tooVolatileRatio: new BN(11), // allow 11x change
+			},
+		};
+
+		await driftClient.updateOracleGuardRails(oracleGuardRails);
+
+		const txSig = await liquidatorDriftClient.liquidatePerp(
+			await driftClient.getUserAccountPublicKey(),
+			driftClient.getUserAccount(),
+			0,
+			new BN(175).mul(BASE_PRECISION).div(new BN(10)).divn(numMkts),
+			undefined,
+			{
+				computeUnits: 300_000,
+			}
+		);
+
+		await printTxLogs(connection, txSig);
+
+		const cus = (
+			await findComputeUnitConsumption(
+				driftClient.program.programId,
+				driftClient.connection,
+				txSig
+			)
+		)[0];
+		console.log(cus);
+	});
+});