+++ CHANGELOG.md
@@ -9,11 +9,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: allow settle pnl and spot fills via match when utilization is 100% ([#525](https://github.com/drift-labs/protocol-v2/pull/525)) 
+- program: new update_perp_bid_ask_twap ix ([#548](https://github.com/drift-labs/protocol-v2/pull/548))
+- program: dont check price bands for place order ([#556](https://github.com/drift-labs/protocol-v2/pull/556))
+
+### Fixes
+
+### Breaking
+
+## [2.36.0] - 2023-07-26
+
+### Features
+
 - program: usdt oracle uses pyth stablecoin oracle source
 
 ### Fixes
 
-- ts-sdk: fix abs for dustBaseAssetValue in getPerpPositionWithLPSettle (#543)
+- program: add buffer before limit tif can be expired ([#551](https://github.com/drift-labs/protocol-v2/pull/551))
+- ts-sdk: fix abs for dustBaseAssetValue in getPerpPositionWithLPSettle ([#543](https://github.com/drift-labs/protocol-v2/pull/543))
+- program: add a fixed buffer margin requirement for lp_shares ([#546](https://github.com/drift-labs/protocol-v2/pull/546))
+- program: use fill margin type in fulfill_spot_order
+- ts-sdk: add buffer to max leverage for LP contributions
 
 ### Breaking
 
@@ -29,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - program: allow up to 10000 users
 
 ### Fixes
+
 - program: if taker increases free colalteral, check maintenance health ([#538](https://github.com/drift-labs/protocol-v2/pull/538))
 - program: improve bid/ask twap update for infrequent trading ([#529](https://github.com/drift-labs/protocol-v2/pull/529))
 - sdk: simplify, mirror contract, and write tests for predicting funding rate function ([#529](https://github.com/drift-labs/protocol-v2/pull/529))